Decreased Stone heading for right path to -13 degrees
After decreasing the tolerance to +/- 1 degree, we realized that the robot had been consistently undershooting its target of -15 degrees but could now end closer to it. It was overshooting before, so we accounted for this by decreasing the heading by two degrees instead of just the one degree necessary to balance out the tolerance changes.